取消bop前置
