+++ build.gradle
@@ -145,8 +145,8 @@ dependencies {
         testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
     }
 
-    implementation rfg.deobf(files("libs/Traverse-1.12.2-1.6.0-69.jar"))
-    implementation rfg.deobf(files("libs/BiomesOPlenty-1.12.2-7.0.1.2445-universal.jar"))
+    compileOnly rfg.deobf(files("libs/BiomesOPlenty-1.12.2-7.0.1.2445-universal.jar"))
+
     implementation rfg.deobf(files("libs/RoughlyEnoughIDs-2.2.2.jar"))
 
 }
