+++ gradle.properties
@@ -15,7 +15,7 @@ show_testing_output = false
 
 # Mod Information
 # HIGHLY RECOMMEND complying with SemVer for mod_version: https://semver.org/
-mod_version = 1.0.1
+mod_version = 1.0.2
 root_package = rtg
 mod_id = rtgc
 mod_name = Realistic-Terrain-Generation-Community
