+++ src/main/java/rtg/LateMixin.java
@@ -1,5 +1,6 @@
 package rtg;
 
+import rtg.compat.ModCompat;
 import zone.rong.mixinbooter.ILateMixinLoader;
 
 import java.util.Collections;
@@ -11,5 +12,11 @@ public class LateMixin implements ILateMixinLoader {
     public List<String> getMixinConfigs() {
         return Collections.singletonList("mixins.rtg_late.json");
     }
-
+    @Override
+    public boolean shouldMixinConfigQueue(String mixinConfig) {
+        if ("mixins.rtg_late.json".equals(mixinConfig)) {
+            return ModCompat.Mods.biomesoplenty.isLoaded();
+        }
+        return true;
+    }
 }
